requirements.txt
eve         == 1.1
flask-cors  == 3.0.8
gunicorn    == 20.0.4
PyYAML      == 5.3
